requirements.txt
console_tea==0.0.1
django==3.1
django-tea==0.0.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
traktor==0.0.3

